requirements.txt
bootstrap4==0.1.0
boto3==1.14.51
botocore==1.17.51
Django==3.0.4
django-bootstrap4==1.1.1
dj-database-url==0.5.0
django-storages==1.9.1
gunicorn==20.0.4
Pillow==7.0.0
pipenv==2020.6.2
psycopg2==2.8.5
s3transfer==0.3.3
whitenoise==5.0.1
